Create plan and pseudocode for display subproblems
Create preventIncorrectExpression()

Currently, user is able to add any button except clear, delete, and '='
to the display in any order they choose. Clicking equals after the
expression either shows undefined or NaN in the display.

Create a plan to prevent the user from making an error in the
expression. Write pseudocode for code clarity. Translate pseudocode to
code.