Add index2.js and changed display font size
font size was too small.

Add index2.js as a secondary and preferred method for how
the calculator should work. The file 'index.js' had many
problems that seemed to stem from using the string from
the display for every operation.

The file index2.js uses an expression array which makes the
code more simple to understand, and maintain. With this
method, it was easier to make validations for each entry in
the expression so that users aren't able to make a error.