Altered the approach for index.js
First, index.js used a string from the display.value in order to get
the 1st and 2nd number plus the operator. Then it calculated the
result of that. This approach was limited but the number of numbers
and operators that could be used in the expression. Secondly,
index.js had no way of restricting the user from making an invalid
expression.

Currently, the user cannot have the operator at an incorrect
position. The user cannot have multiple decimals for each number.
The user can enter any amount of numbers for each number in
the expression. Finally, now the user can enter any amount of
numbers and operators in the expression.