Add divide by 0, and calculate functionality
calculateExpression() checks if the user divided by 0. If so,
a unique message will appear.

calculateExpression() gives the equals button functionality.
This functions purpose is to evaluate the expression and change
the display.

This function checks how many decimal places the result will have.
If the result has 0, 1, 2 or more decimal places, it changes the
display accordingly.