Add keyboard support, and format error functions
Using the keyboard -- Esc, Backspace, and Enter can clear, delete
the last value, and calculate the expression respectfully.

In addition, the validation check for each input is the same as the
click event. However, there is a minor error when using the '*'
operator at the beginning of the expression in the display.

A new format error message has been added for when the user has a
single decimal as an item in their expression.